Split durable adapter test layers
