+++ packages/xstate-inngest/README.md
@@ -36,3 +36,14 @@ an inline `step.sleep()` would block intervening events and cancellation.
 
 Use `createInngestAdapter()` with XState's lower-level `createDurable()` when
 you need to assemble the transition loop yourself.
+
+## Tests
+
+<!-- test layers from vitest configs and package.json#scripts -->
+
+Fast context-contract tests run in the normal monorepo suite. SDK integration
+tests are opt-in:
+
+```sh
+pnpm --filter @xstate/inngest test:integration
+```
