+++ packages/xstate-inngest/vitest.config.mts
@@ -3,7 +3,7 @@ import { defineProject } from 'vitest/config';
 export default defineProject({
   test: {
     name: '@xstate/inngest',
-    include: ['test/**/*.test.ts'],
+    include: ['test/contract/**/*.test.ts'],
     globals: true,
     environment: 'node'
   }
