+++ packages/xstate-inngest/vitest.integration.config.mts
@@ -1,9 +1,10 @@
 import { defineProject } from 'vitest/config';
 
 export default defineProject({
+  root: import.meta.dirname,
   test: {
-    name: '@xstate/rivet',
-    include: ['test/contract/**/*.test.ts'],
+    name: '@xstate/inngest-integration',
+    include: ['test/integration/**/*.test.ts'],
     globals: true,
     environment: 'node'
   }
