+++ packages/xstate-rivet/package.json
@@ -32,7 +32,9 @@
     "type": "git",
     "url": "git+https://github.com/statelyai/xstate.git"
   },
-  "scripts": {},
+  "scripts": {
+    "test:integration": "vitest run --config vitest.integration.config.mts"
+  },
   "peerDependencies": {
     "rivetkit": "^2.3.10",
     "xstate": "workspace:^"
